Refactor: split internal/database/db.go into per-concern files (#64)
Splits the 910-line monolithic db.go into five focused files within the
same package: db.go (core struct/lifecycle/migrations), refresh.go,
channels.go, airings.go, search.go, and icons.go. No behavioral changes.
